fs: sync: Avoid calling fdget without fdput, set proper permissions for fsync_enabled fixes "[    0.359222] [    0.359218]@7 Attribute fsync_enabled: Invalid permissions 0755"
When adding fsync support, we check for fsync_enabled() in several
cases but it appears that we should fdput() after fdget() but the
current code just check for fsync_enabled and directly return in
some cases after calling fdget().

Fix it by checking fsync_enabled first and move the f initialization
after the check. Also remove some unnecessary fsync_enabled checks as
this will be checked later in do_fsync().